custom_api: enhance employee check-in handling to prevent errors by adding checks for attendance existence before cancellation and deletion, and ensure database commit after processing

+++ custom_api/api.py
@@ -2412,7 +2412,10 @@ def create_doc_and_multi_payment(reference_doc,multi_payment,submit=False):
 
 
 def employee_checkin_on_update(doc, method):
-    attendance = frappe.get_doc("Attendance", doc.attendance)
-    attendance.cancel()
-    attendance.delete()
+    if doc.attendance != None and doc.attendance != "":
+        attendance = frappe.get_doc("Attendance", doc.attendance)
+        attendance.cancel()
+        attendance.delete()
     create_employee_attendance(doc.company)
+
+    frappe.db.commit()
